# Conflicts:
#	local.env
#	src/main/java/edu/codegym/toam/config/WebSecurityConfig.java
#	src/main/java/edu/codegym/toam/controller/CommentRestController.java
#	src/main/java/edu/codegym/toam/controller/DistrictRestController.java
#	src/main/java/edu/codegym/toam/controller/PropertiesRestController.java
#	src/main/java/edu/codegym/toam/model/Comments.java
#	src/main/java/edu/codegym/toam/model/Properties.java
#	src/main/java/edu/codegym/toam/model/PropertyStatus.java
#	src/main/java/edu/codegym/toam/repository/DistrictRepository.java
#	src/main/java/edu/codegym/toam/repository/PropertiesRepository.java
#	src/main/java/edu/codegym/toam/service/comments/CommentsService.java
#	src/main/java/edu/codegym/toam/service/comments/ICommentsService.java
#	src/main/java/edu/codegym/toam/service/district/DistrictService.java
#	src/main/java/edu/codegym/toam/service/district/IDistrictService.java
#	src/main/java/edu/codegym/toam/service/properties/IPropertiesService.java
#	src/main/java/edu/codegym/toam/service/properties/PropertiesService.java

+++ src/main/java/edu/codegym/toam/controller/AccountRestController.java
@@ -8,6 +8,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.NotActiveException;
+
 @RestController
 @RequestMapping("api/account")
 @CrossOrigin(origins = "*")
@@ -22,6 +24,7 @@ public class AccountRestController {
         return ResponseEntity.ok(this.accountService.findAll());
     }
 
+    //Lấy danh sách tất cả những thằng chủ nhà
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @GetMapping("/host")
     public ResponseEntity<Iterable<Account>> getHost() {
@@ -47,7 +50,6 @@ public class AccountRestController {
     }
 
     @PostMapping()
-    @PreAuthorize("hasAnyRole('ROLE_RENTER','ROLE_ADMIN','ROLE_HOST')")
     public ResponseEntity<Account> createAccount(@RequestBody Account account) {
         try {
             return ResponseEntity.ok(this.accountService.create(account));
@@ -66,12 +68,31 @@ public class AccountRestController {
         }
     }
 
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> removeAccountById(@PathVariable Long id) {
+
+    //    Chuyển đổi trạng thái account (từ block sang không block và ngược lại)
+    @PostMapping("/edit/{accountId}")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN','ROLE_RENTER','ROLE_HOST')")
+    public ResponseEntity<Account> changeAccountStatus(@PathVariable Long accountId) {
         try {
-            this.accountService.removeById(id);
+            this.accountService.changeAccountStatus(accountId);
             return new ResponseEntity<>(HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+        }
+    }
+
+    //    Xóa account điều kiện ràng buộc là ko có bất kỳ hóa đơn nào
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @GetMapping("delete/{id}")
+    public ResponseEntity<String> removePropertyById(@PathVariable Long id) {
+        try {
+            if (this.accountService.findById(id) == null) throw new Exception();
+            if (!this.accountService.checkAccountConstraint(id)) {
+                this.accountService.removeById(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else throw new NotActiveException();
+        } catch (NotActiveException nae) {
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
